Fix cmd options plus replace for custom windows search

+++ lua/config/options.lua
@@ -5,15 +5,17 @@
 -- vim.opt.shiftwidth = 4
 -- vim.opt.expandtab = true
 vim.diagnostic.config({ virtual_text = true, virtual_lines = { current_line = true } })
-if vim.fn.executable("pwsh") == 1 then
-  vim.opt.shell = "pwsh"
-else
-  vim.opt.shell = "powershell"
-end
+if jit.os == "Windows" then
+  if vim.fn.executable("pwsh") == 1 then
+    vim.opt.shell = "pwsh"
+  else
+    vim.opt.shell = "powershell"
+  end
 
-vim.opt.shellcmdflag =
-  [[-NoLogo -NonInteractive -ExecutionPolicy RemoteSigned -Command [Console]::InputEncoding=[Console]::OutputEncoding=[System.Text.UTF8Encoding]::new();$PSDefaultParameterValues['Out-File:Encoding']='utf8';$PSStyle.OutputRendering='plaintext';Remove-Alias -Force -ErrorAction SilentlyContinue tee;]]
-vim.opt.shellredir = [[2>&1 | %{$_.ToString()} | Out-File %s; exit $LastExitCode]]
-vim.opt.shellpipe = [[2>&1 | %{$_.ToString()} | tee %s; exit $LastExitCode]]
-vim.opt.shellquote = ""
-vim.opt.shellxquote = ""
+  vim.opt.shellcmdflag =
+    [[-NoLogo -NonInteractive -ExecutionPolicy RemoteSigned -Command [Console]::InputEncoding=[Console]::OutputEncoding=[System.Text.UTF8Encoding]::new();$PSDefaultParameterValues['Out-File:Encoding']='utf8';$PSStyle.OutputRendering='plaintext';Remove-Alias -Force -ErrorAction SilentlyContinue tee;]]
+  vim.opt.shellredir = [[2>&1 | %{$_.ToString()} | Out-File %s; exit $LastExitCode]]
+  vim.opt.shellpipe = [[2>&1 | %{$_.ToString()} | tee %s; exit $LastExitCode]]
+  vim.opt.shellquote = ""
+  vim.opt.shellxquote = ""
+end
